dom/media/test/test_can_play_type_webm.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=566245
-->
<head>
  <title>Test for Bug 566245: WebM backend</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="text/javascript" src="manifest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=566245">Mozill
a Bug 566245</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<video id="v"></video>

<pre id="test">
<script src="can_play_type_webm.js"></script>
<script>
  async function runTest() {
    try {
      await check_webm(document.getElementById('v'), true);
      mediaTestCleanup();
    } catch (e) {
      info("Exception " + e.message);
      ok(false, "Threw exception " + e.message);
    }
    SimpleTest.finish();
  }
  SimpleTest.waitForExplicitFinish();
  addLoadEvent(runTest);
</script>
</pre>
</body>
</html>
